requirements.txt
transformers
optimum
accelerate
fastapi
uvicorn
sse-starlette
pydantic
firebase-admin
intel_extension_for_pytorch